AI-HRMS/mcp-server/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.0
python-json-logger==2.0.7
